Support back button Change animation to fade Insert confirm dialogs for save and delete Trivial looking changes

+++ bundle.js
@@ -68,10 +68,12 @@ window.onload = function () {
   $("#btn-start1").on('click', start1);
   $("#btn-start2").on('click', start2);
   $("#btn-next").on('click', nextStatement);
+  $("#btn-back").on('click', backStatement);
+  $("#btn-back-to-progress").on('click', backToProgress);
   $("#btn-history").on('click', showHistory);
   $("#btn-init").on('click', initialize);
   $("#btn-save").on('click', saveResult);
-  $("#btn-delete-all").on('click', deleteAll);
+  $(".btn-delete-all").on('click', deleteAll);
 }
 
 function initialize() {
@@ -98,6 +100,7 @@ function startCommon(num, statements) {
 }
 
 function showCountdown() {
+  $("#progress-bar").width("0%");
   display("countdown");
   $("#countdown-sec").text("3");
   setTimeout(() => {
@@ -109,13 +112,12 @@ function showCountdown() {
         nextStatement();
       }, 1000);
     }, 1000);
-  }, 1000);
-  nextStatement();
+  }, 1000 + 500);
 }
 
 function display(block, speed) {
   if (speed === undefined) {
-    speed = 300;
+    speed = 500;
   }
   const contents = [
     "start",
@@ -127,7 +129,7 @@ function display(block, speed) {
   contents.forEach(c => {
     $(`#${c}`).hide();
   });
-  $(`#${block}`).show(speed);
+  $(`#${block}`).show('fade', '', speed);
   document.getElementById("goto-init").style.display = "start" == block ? "none" : "block";
   document.getElementById("goto-history").style.display = "history" == block ? "none" : "block";
 }
@@ -137,6 +139,7 @@ function nextStatement() {
     showResult();
     return;
   }
+  document.getElementById("btn-back").style.display = state.progress > 0 ? "block" : "none";
 
   $("#statement").text(state.statements[state.progress]);
 
@@ -145,11 +148,22 @@ function nextStatement() {
   $("#progress-bar").width(`${percentage}%`);
 }
 
+function backStatement() {
+  state.progress -= 2;
+  nextStatement();
+}
+
 function showResult() {
   state.endTime = new Date();
   display("result");
   $("#card-title").text(state.num);
   $("#elapsed").text(formatElapsedTime(state.endTime - state.startTime));
+}
+
+function backToProgress() {
+  display("progress", 0);
+  state.progress = state.statements.length - 1;
+  nextStatement();
 }
 
 function showHistory(speed) {
@@ -212,23 +226,27 @@ function showHistoryItems(tag, historyItems) {
     tag.append(
       `<div class="col s7 ${h.elapsedMs == minElapsedMs ? "green lighten-3" : ""}" style="margin-left: 10pt; white-space: nowrap">${h.date}</div>
       <div class="col s3 ${h.elapsedMs == minElapsedMs ? "green lighten-3 left" : ""}" style="white-space: nowrap">${formatElapsedTime(h.elapsedMs)}</div>
-      <div class="col s1 delete-this" data-date="${h.date}"><i class="material-icons">clear</i></div>`
+      <div class="col s1 delete-this" data-date="${h.date}"><i class="material-icons" style="color: grey; font-size: 1em">clear</i></div>`
     );
   });
   $(".delete-this", tag).on("click", function () {
-    deleteHistoryItem($(this).data("date"));
-    showHistory(0);
+    if (confirm(`${$(this).data("date")} の記録を削除しますか？`)) {
+      deleteHistoryItem($(this).data("date"));
+      showHistory(0);
+    }
   });
 }
 
 function saveResult() {
-  const result = JSON.parse(localStorage.getItem(state.num) || "[]");
-  result.unshift({
-    date: state.startTime.toLocaleString(),
-    elapsedMs: state.endTime - state.startTime
-  });
-  localStorage.setItem(state.num, JSON.stringify(result));
-  showHistory();
+  if (confirm(`${state.num} の記録 ${formatElapsedTime(state.endTime - state.startTime)} を保存しますか？`)) {
+    const result = JSON.parse(localStorage.getItem(state.num) || "[]");
+    result.unshift({
+      date: state.startTime.toLocaleString(),
+      elapsedMs: state.endTime - state.startTime
+    });
+    localStorage.setItem(state.num, JSON.stringify(result));
+    showHistory();
+  }
 }
 
 function deleteHistoryItem(date) {
@@ -247,9 +265,9 @@ function deleteHistoryItem(date) {
 }
 
 function deleteAll() {
-  if (confirm("記録を全部消していいですか？")) {
-    localStorage.removeItem("No. 1");
-    localStorage.removeItem("No. 2");
+  const num = $(this).data("num");
+  if (confirm(`${num}の記録を全部消していいですか？`)) {
+    localStorage.removeItem(num);
     showHistory(0);
   }
 }
